requirements.txt
asgiref==3.7.2
Django==4.2.6
djangorestframework==3.14.0
psycopg==3.1.12
psycopg2-binary==2.9.9
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.8.0
